Make the client a seperate crate
