+++ experimental/uefi/runtime/src/remote_attestation.rs
@@ -23,7 +23,7 @@ extern crate alloc;
 
 use alloc::vec::Vec;
 use anyhow::Context;
-use oak_remote_attestation_sessions::server::{SerializeableRequest, SessionState, SessionTracker};
+use oak_remote_attestation_sessions::{SerializeableRequest, SessionState, SessionTracker};
 
 /// Number of sessions that will be kept in memory.
 const SESSIONS_CACHE_SIZE: usize = 10000;
