+++ grpc_unary_attestation/Cargo.toml
@@ -7,9 +7,8 @@ license = "Apache-2.0"
 
 [dependencies]
 oak_remote_attestation = { path = "../remote_attestation/rust/" }
-oak_remote_attestation_sessions = { path = "../remote_attestation_sessions/", features = [
-  "client"
-] }
+oak_remote_attestation_sessions = { path = "../remote_attestation_sessions/" }
+oak_remote_attestation_sessions_client = { path = "../remote_attestation_sessions_client/" }
 anyhow = "*"
 prost = "*"
 prost-types = "*"
