+++ remote_attestation_sessions_client/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "oak_remote_attestation_sessions"
+name = "oak_remote_attestation_sessions_client"
 version = "0.1.0"
 authors = ["Juliette Pretot <julsh@google.com>"]
 edition = "2021"
@@ -13,4 +13,6 @@ rust-crypto = ["oak_remote_attestation/rust-crypto"]
 [dependencies]
 anyhow = { version = "*", default-features = false }
 oak_remote_attestation = { path = "../remote_attestation/rust/", default-features = false }
-lru = "*"
+oak_remote_attestation_sessions = { path = "../remote_attestation_sessions", default-features = false }
+async-trait = "*"
+rand = "*"
